space/requirements.txt
gradio>=5.0,<6.0
fal-client>=0.5,<1.0
pillow>=10.0,<12.0
requests>=2.31,<3.0
